comecei componente de criar post
