+++ modulo3/projeto-labeddit/labeddit/src/components/PostForm/PostForm.js
@@ -1,0 +1,10 @@
+const PostForm = () => {
+    
+      return (
+        <div>
+          PostForm
+        </div>
+      );
+    }
+    
+    export default PostForm;
